+++ Sources/PeripheryKit/Results/OutputFormatter.swift
@@ -37,6 +37,8 @@ extension OutputFormatter {
             "redundantInternalAccessibility"
         case .redundantFilePrivateAccessibility:
             "redundantFilePrivateAccessibility"
+        case .superfluousIgnoreCommand:
+            "superfluousIgnoreCommand"
         }
     }
 
@@ -71,11 +73,11 @@ extension OutputFormatter {
             case let .redundantInternalAccessibility(_, suggestedAccessibility):
                 let accessibilityText = suggestedAccessibility?.rawValue ?? "private/fileprivate"
                 description += "Redundant internal accessibility for \(kindDisplayName) '\(name)' (not used outside of file; can be \(accessibilityText))"
-            // Hint: if we wanted to output the USR for helping build bazel.json, we could also output:
-            // result.declaration.usrs.joined(separator: ", ")
             case let .redundantFilePrivateAccessibility(_, containingTypeName):
                 let context = containingTypeName.map { "only used within \($0)" } ?? "not used outside of file"
                 description += "Redundant fileprivate accessibility for \(kindDisplayName) '\(name)' (\(context); can be private)"
+            case .superfluousIgnoreCommand:
+                description += "Superfluous ignore comment for \(kindDisplayName) '\(name)' (declaration is referenced and should not be ignored)"
             }
         } else {
             description += "Unused"